+++ project/src/App.jsx
@@ -1,0 +1,10 @@
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+    </div>
+  );
+}
+
+export default App;
